Adding leaky_relu fix and oneDNN_primitive_cache fix
